Move from struct to ctypes
The Header class is split in two parts. The first part is a structure
that only represents the magic, CRC and type. The second part is moved
to the PAK class that now carries the sections, partitions and is64
attributes.

get_section_count() is removed. Instead of reading all sections twice,
once to get the amount and again to actually read them, we read sections
until we get to the first partition and then seek back a section's size
to read the partitions.